+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## 更新日志
 
+### 2.0.0
+
+- 修改 method默认值由'get'改为'post'
+
+- 修改 accept默认值由{extensions: '', mimeTypes: ''}改为null
+
+- 修改 sendAsBinary默认值由true改为false
+
+- 修改 chunked默认值由true改为false
+
+- 新增 新增compress属性，默认值为false
+
 ### 1.1.0
 
 *2018-02-06*